feat: add new hooks and not-found routes, change links hrefs
