+++ app/hooks/[category]/page.tsx
