+++ app/not-found.tsx
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  <>Hi - NotFound</>;
+}
